Split accept module; move server into its own module
