+++ linkerd/app/inbound/src/lib.rs
@@ -6,6 +6,7 @@
 #![deny(warnings, rust_2018_idioms)]
 #![forbid(unsafe_code)]
 
+mod accept;
 mod detect;
 pub mod direct;
 mod http;
@@ -16,14 +17,14 @@ pub(crate) mod test_util;
 
 pub use self::port_policies::PortPolicies;
 use linkerd_app_core::{
-    config::{ConnectConfig, ProxyConfig, ServerConfig},
-    drain, io, metrics, profiles,
+    config::{ConnectConfig, ProxyConfig},
+    drain, io, metrics,
     proxy::tcp,
-    serve, svc,
-    transport::{self, listen::Bind, ClientAddr, Local, OrigDstAddr, Remote, ServerAddr},
+    svc,
+    transport::{self, Remote, ServerAddr},
     Error, NameMatch, ProxyRuntime,
 };
-use std::{fmt::Debug, future::Future, time::Duration};
+use std::{fmt::Debug, time::Duration};
 use tracing::debug_span;
 
 #[derive(Clone, Debug)]
@@ -39,11 +40,6 @@ pub struct Inbound<S> {
     config: Config,
     runtime: ProxyRuntime,
     stack: svc::Stack<S>,
-}
-
-#[derive(Copy, Clone, Debug)]
-struct TcpEndpoint {
-    port: u16,
 }
 
 // === impl Inbound ===
@@ -134,73 +130,6 @@ impl Inbound<()> {
                 })
         })
     }
-
-    pub fn serve<B, G, GSvc, P>(
-        self,
-        bind: B,
-        profiles: P,
-        gateway: G,
-    ) -> (Local<ServerAddr>, impl Future<Output = ()> + Send)
-    where
-        B: Bind<ServerConfig>,
-        B::Addrs: svc::Param<Remote<ClientAddr>>
-            + svc::Param<Local<ServerAddr>>
-            + svc::Param<OrigDstAddr>,
-        G: svc::NewService<direct::GatewayConnection, Service = GSvc>,
-        G: Clone + Send + Sync + Unpin + 'static,
-        GSvc: svc::Service<direct::GatewayIo<io::ScopedIo<B::Io>>, Response = ()> + Send + 'static,
-        GSvc::Error: Into<Error>,
-        GSvc::Future: Send,
-        P: profiles::GetProfile<profiles::LookupAddr> + Clone + Send + Sync + Unpin + 'static,
-        P::Error: Send,
-        P::Future: Send,
-    {
-        let (Local(ServerAddr(la)), listen) = bind
-            .bind(&self.config.proxy.server)
-            .expect("Failed to bind inbound listener");
-
-        let serve = async move {
-            let shutdown = self.runtime.drain.clone().signaled();
-
-            // Handles connections to ports that can't be determined to be HTTP.
-            let forward = self
-                .clone()
-                .into_tcp_connect(la.port())
-                .push_tcp_forward()
-                .into_stack()
-                .push_map_target(TcpEndpoint::from_param)
-                .instrument(|_: &_| debug_span!("tcp"))
-                .into_inner();
-
-            // Handles connections that target the inbound proxy port.
-            let direct = self
-                .clone()
-                .into_tcp_connect(la.port())
-                .push_tcp_forward()
-                .map_stack(|_, _, s| s.push_map_target(TcpEndpoint::from_param))
-                .push_direct(gateway)
-                .into_stack()
-                .instrument(|_: &_| debug_span!("direct"))
-                .into_inner();
-
-            // Handles HTTP connections.
-            let http = self
-                .into_tcp_connect(la.port())
-                .push_http_router(profiles)
-                .push_http_server();
-
-            // Determines how to handle an inbound connection, dispatching it to the appropriate
-            // stack.
-            let server = http
-                .push_detect(forward)
-                .push_accept(la.port(), direct)
-                .into_inner();
-
-            serve::serve(listen, server, shutdown).await
-        };
-
-        (Local(ServerAddr(la)), serve)
-    }
 }
 
 impl<S> Inbound<S> {
@@ -245,24 +174,4 @@ impl<S> Inbound<S> {
 
 fn stack_labels(proto: &'static str, name: &'static str) -> metrics::StackLabels {
     metrics::StackLabels::inbound(proto, name)
-}
-
-// === impl TcpEndpoint ===
-
-impl TcpEndpoint {
-    pub fn from_param<T: svc::Param<u16>>(t: T) -> Self {
-        Self { port: t.param() }
-    }
-}
-
-impl svc::Param<u16> for TcpEndpoint {
-    fn param(&self) -> u16 {
-        self.port
-    }
-}
-
-impl svc::Param<transport::labels::Key> for TcpEndpoint {
-    fn param(&self) -> transport::labels::Key {
-        transport::labels::Key::InboundConnect
-    }
 }
